update: working LZW implementation with 4096 chunk size.
- The issue was that the hash function was performing the
  bitwise AND operation with `0x7FFF` which is 15 bits, although
  the map capacity was set to 65536. This meant that we were not
  utilizing the complete map, due to which there were problems. This
  has been fixed in this commit.
- This commit also contains the corrected testbench for Vitis HLS.

+++ project/Encoder/common.h
@@ -34,7 +34,7 @@ int64_t dedup(string sha_fingerprint);
 //          uint32_t *code_length);
 
 void lzw(unsigned char *chunk, uint32_t start_idx, uint32_t end_idx,
-         uint16_t *lzw_codes, uint32_t *code_length, uint8_t *failure,
+         uint32_t *lzw_codes, uint32_t *code_length, uint8_t *failure,
 		 unsigned int *associative_mem);
 
 #endif
